+++ flake8_annotations/__init__.py
@@ -1,6 +1,5 @@
 import sys
 from itertools import zip_longest
-from pathlib import Path
 from typing import List, Union
 
 from flake8_annotations.enums import AnnotationType, ClassDecoratorType, FunctionType
@@ -10,10 +9,12 @@ from flake8_annotations.enums import AnnotationType, ClassDecoratorType, Functio
 if sys.version_info >= (3, 8):
     import ast
     from ast import Ellipsis as ast_Ellipsis
+
     PY_GTE_38 = True
 else:
     from typed_ast import ast3 as ast
     from typed_ast.ast3 import Ellipsis as ast_Ellipsis
+
     PY_GTE_38 = False
 
 __version__ = "1.1.1"
@@ -260,7 +261,7 @@ class Function:
 
     @staticmethod
     def get_class_decorator_type(
-        function_node: AST_FUNCTION_TYPES
+        function_node: AST_FUNCTION_TYPES,
     ) -> Union[ClassDecoratorType, None]:
         """
         Get the class method's decorator type from its function node.
@@ -333,23 +334,3 @@ class FunctionVisitor(ast.NodeVisitor):
         # Use ast.NodeVisitor.generic_visit to start down the nested method chain
         for sub_node in node.body:
             self.generic_visit(sub_node)
-
-    @classmethod
-    def parse_file(cls, src_filepath: Path) -> "FunctionVisitor":  # Need to quote for 3.6 compat
-        """Return a parsed AST for the provided Python source file."""
-        with src_filepath.open("r", encoding="utf-8") as f:
-            src = f.read()
-
-        if PY_GTE_38:
-            # Built-in ast requires a flag to parse type comments
-            tree = ast.parse(src, type_comments=True)
-        else:
-            # typed-ast will implicitly parse type comments
-            tree = ast.parse(src)
-
-        lines = src.splitlines()
-
-        visitor = cls(lines)
-        visitor.visit(tree)
-
-        return visitor